Add user resolvers to separate folder
