+++ graphql/resolvers/users/index.js
@@ -1,0 +1,7 @@
+const { userRegister } = require("./register");
+
+module.exports = {
+  Mutation: {
+    register: userRegister,
+  },
+};
